Append option now added to job id
The job id in template scripts is used for the job name as well as the job output file. This edit adds the append string to the job id to prevent name clashes in those areas as well.

Additional minor change: number of parameters found in the template file is now correctly reported.